dnsproxy: Extract and export ipFamily functions
This will be useful for the upcoming commit to use in the datapath /
iptables packages for extracting shared state from iptables rules
related to the DNS proxy.
